Add tests for writing to real storage; Add verify job
