+++ .github/workflows/build.yml
@@ -92,10 +92,11 @@ jobs:
           - python_deps_ids: [""]
             matrix_override: ${{fromJson(needs.common_config.outputs.linux_matrix)}}
     name: Seed Real Storage for 3.${{matrix.python3}} Linux
-    uses: ./.github/workflows/real_storage_seed.yml
+    uses: ./.github/workflows/real_storage.yml
     secrets: inherit
     permissions: {packages: write}
     with:
+      job_type: seed
       python3: ${{matrix.python3}}
       matrix: ${{toJson(matrix.matrix_override)}}
       python_deps_ids: ${{toJson(matrix.python_deps_ids)}}
@@ -112,10 +113,11 @@ jobs:
           - python_deps_ids: [""]
             matrix_override: ${{fromJson(needs.common_config.outputs.windows_matrix)}}
     name: Seed Real Storage for 3.${{matrix.python3}} Windows
-    uses: ./.github/workflows/real_storage_seed.yml
+    uses: ./.github/workflows/real_storage.yml
     secrets: inherit
     permissions: {packages: write}
     with:
+      job_type: seed
       python3: ${{matrix.python3}}
       matrix: ${{toJson(matrix.matrix_override)}}
       python_deps_ids: ${{toJson(matrix.python_deps_ids)}}
@@ -228,6 +230,49 @@ jobs:
       cmake_preset_type: ${{needs.common_config.outputs.cmake_preset_type_resolved}}
       matrix: ${{needs.common_config.outputs.windows_matrix}}
       real_storage: ${{ needs.common_config.outputs.real_storage }}
+
+  real_storage_verify_linux:
+    # Then use the cached compilation artifacts to build other python versions concurrently in cibuildwheels
+    needs: [common_config, follower-linux]
+    if: needs.common_config.outputs.real_storage
+    strategy:
+      fail-fast: false
+      matrix:
+        python3: ${{fromJson(vars.LINUX_PYTHON_VERSIONS || '[6, 7, 8, 9, 10, 11]')}}
+        include:
+          - python_deps_ids: [""]
+            matrix_override: ${{fromJson(needs.common_config.outputs.linux_matrix)}}
+    name: Verify Real Storage for 3.${{matrix.python3}} Linux
+    uses: ./.github/workflows/real_storage.yml
+    secrets: inherit
+    permissions: {packages: write}
+    with:
+      job_type: verify
+      python3: ${{matrix.python3}}
+      matrix: ${{toJson(matrix.matrix_override)}}
+      python_deps_ids: ${{toJson(matrix.python_deps_ids)}}
+
+  real_storage_verify_windows:
+    # Then use the cached compilation artifacts to build other python versions concurrently in cibuildwheels
+    needs: [common_config, follower-windows]
+    if: needs.common_config.outputs.real_storage
+    strategy:
+      fail-fast: false
+      matrix:
+        python3: ${{fromJson(vars.WINDOWS_PYTHON_VERSIONS || '[7, 8, 9, 10, 11]')}}
+        include:
+          - python_deps_ids: [""]
+            matrix_override: ${{fromJson(needs.common_config.outputs.windows_matrix)}}
+    name: Verify Real Storage for 3.${{matrix.python3}} Windows
+    uses: ./.github/workflows/real_storage.yml
+    secrets: inherit
+    permissions: {packages: write}
+    with:
+      job_type: verify
+      python3: ${{matrix.python3}}
+      matrix: ${{toJson(matrix.matrix_override)}}
+      python_deps_ids: ${{toJson(matrix.python_deps_ids)}}
+    
 
   can_merge:
     needs: [leader-cpp-test-linux, leader-cpp-test-windows, follower-linux, follower-windows]
