+++ .github/workflows/build_steps.yml
@@ -270,6 +270,7 @@ jobs:
           echo "ARCTICDB_REAL_S3_CLEAR=1" >> $GITHUB_ENV
           echo "ARCTICDB_REAL_S3_ACCESS_KEY=${{ secrets.AWS_S3_ACCESS_KEY }}" >> $GITHUB_ENV
           echo "ARCTICDB_REAL_S3_SECRET_KEY=${{ secrets.AWS_S3_SECRET_KEY }}" >> $GITHUB_ENV
+          echo "ARCTICDB_REAL_STORAGE_LIB_NAME=${{ github.ref_name }}_${{matrix.os}}_3.${{inputs.python3}}" >> $GITHUB_ENV
 
       - name: Run test
         run: build_tooling/parallel_test.sh tests/${{matrix.type}}
